dataviz::figure::figuretypes::groupbarchart
Struct GroupBarChart

pub struct GroupBarChart {
    pub datasets: Vec<BarDataset>,
    pub title: String,
    pub x_label: String,
    pub y_label: String,
    pub orientation: Orientation,
    pub config: FigureConfig,
}

A grouped bar chart representation, supporting horizontal and vertical orientations.

Fields
datasets: Vec<BarDataset>

A collection of datasets to be visualized in the grouped bar chart.

title: String

Title of the bar chart.

x_label: String

Label for the X-axis.

y_label: String

Label for the Y-axis.

orientation: Orientation

Orientation of the bar chart (Horizontal or Vertical).

config: FigureConfig

Configuration settings for rendering the chart (e.g., colors, fonts, grid).

Implementations

impl GroupBarChart

pub fn new( title: &str, x_label: &str, y_label: &str, orientation: Orientation, config: FigureConfig, ) -> Self

Creates a new GroupBarChart instance with the specified title, labels, orientation, and configuration.

Parameters

title: The title of the bar chart.

x_label: The label for the X-axis.

y_label: The label for the Y-axis.

orientation: The orientation of the bar chart (Horizontal or Vertical).

config: The FigureConfig containing appearance and behavior settings for the chart.

Returns

A new GroupBarChart instance with an empty dataset.

Example

use crate::figure::configuration::figureconfig::FigureConfig;
use crate::figure::groupbarchart::GroupBarChart;
use crate::figure::utilities::orientation::Orientation;

let config = FigureConfig::default();
let bar_chart = GroupBarChart::new("Sales Chart", "Year", "Revenue", Orientation::Vertical, config);

pub fn add_dataset(&mut self, dataset: BarDataset)

Adds a dataset to the grouped bar chart.

Parameters

dataset: The BarDataset to be added to the chart.

Example

use crate::figure::datasets::bardataset::BarDataset;

let dataset = BarDataset::new("Company A", [255, 0, 0]);
bar_chart.add_dataset(dataset);

pub fn draw_horizontal(&self, canvas: &mut PixelCanvas)

Draws a horizontal grouped bar chart.

Parameters

canvas: The PixelCanvas to render the bar chart on.

Details

This method draws a horizontally oriented bar chart with grouped bars. It uses the FigureConfig settings for appearance and adjusts the canvas accordingly.

Example

bar_chart.draw_horizontal(&mut canvas);

pub fn draw_vertical(&self, canvas: &mut PixelCanvas)

Draws a vertical grouped bar chart.

Parameters

canvas: The PixelCanvas to render the bar chart on.

Details

This method draws a vertically oriented bar chart with grouped bars. It uses the FigureConfig settings for appearance and adjusts the canvas accordingly.

Example

bar_chart.draw_vertical(&mut canvas);